requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.30.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
pillow>=8.0.0
opencv-python>=4.5.0
albumentations>=1.3.0
fastapi>=0.100.0
uvicorn>=0.22.0
gradio>=3.40.0
python-multipart>=0.0.6
aiofiles>=23.0.0
plotly>=5.15.0
jupyter>=1.0.0
ipywidgets>=8.0.0
tqdm>=4.65.0
